refactor the whole structure
