+++ Cargo.toml
@@ -23,22 +23,24 @@ aes-gcm = { version = "0.10.3", default-features = false, features = ["aes", "al
 chacha20poly1305 = { version = "0.10.1", default-features = false }
 crypto-common = { version = "0.1.6", default-features = false }
 der = { version = "0.7.9", default-features = false }
+derive_more = { version = "1.0.0", default-features = false, features = ["display", "debug", "deref", "from", "into", "error"] }
 digest = { version = "0.10.7", default-features = false }
-ecdsa = { version = "0.16.8", default-features = false, features = ["alloc"] }
+ecdsa = { version = "0.16.9", default-features = false, features = ["alloc"] }
 ed25519-dalek = { version = "2", default-features = false, features = ["pkcs8"] }
 hmac = { version = "0.12.1", default-features = false }
 p256 = { version = "0.13.2", default-features = false, features = ["pem", "ecdsa", "ecdh"] }
 p384 = { version = "0.13.0", default-features = false, features = ["pem", "ecdsa", "ecdh"] }
-paste = { version = "1.0.14", default-features = false }
+paste = { version = "1.0.15", default-features = false }
+pkcs1 = { version = "0.7.5", default-features = false }
 pkcs8 = { version = "0.10.2", default-features = false, features = ["pem", "pkcs5"] }
-pki-types = { package = "rustls-pki-types", version = "1.0.1", default-features = false }
+pki-types = { package = "rustls-pki-types", version = "1.8.0", default-features = false }
 rand_core = { version = "0.6.4", default-features = false, features = ["getrandom"] }
-rsa = { version = "0.9.2", default-features = false, features = ["sha2"] }
-rustls = { version = "0.23.0", default-features = false }
+rsa = { version = "0.9.6", default-features = false, features = ["sha2"] }
+rustls = { version = "0.23.13", default-features = false }
 sec1 = { version = "0.7.3", default-features = false, features = ["pkcs8", "pem"] }
-sha2 = { version = "0.10.7", default-features = false }
-signature = { version = "2.1.0", default-features = false }
-webpki = { package = "rustls-webpki", version = "0.102.0", default-features = false }
+sha2 = { version = "0.10.8", default-features = false }
+signature = { version = "2.2.0", default-features = false }
+webpki = { package = "rustls-webpki", version = "0.102.8", default-features = false }
 x25519-dalek = { version = "2", default-features = false }
 
 [dev-dependencies]
@@ -53,7 +55,7 @@ tls12 = ["rustls/tls12"]
 # zeroize is another typical that can be turned off
 
 # TODO: go through all of these that what gets exposed re: std error type
-std = ["alloc", "webpki/std", "pki-types/std", "rustls/std", "ed25519-dalek/std"]
+std = ["alloc", "webpki/std", "pki-types/std", "rustls/std", "ed25519-dalek/std", "pkcs1/std"]
 # TODO: go through all of these to ensure to_vec etc. impls are exposed
-alloc = ["webpki/alloc", "pki-types/alloc", "aead/alloc", "ed25519-dalek/alloc"]
+alloc = ["webpki/alloc", "pki-types/alloc", "aead/alloc", "ed25519-dalek/alloc", "pkcs1/alloc"]
 zeroize = ["ed25519-dalek/zeroize", "x25519-dalek/zeroize"]
