+++ src/tls12/schemes.rs
@@ -1,0 +1,16 @@
+use rustls::SignatureScheme;
+
+pub const TLS12_ECDSA_SCHEMES: [SignatureScheme; 4] = [
+    SignatureScheme::ECDSA_NISTP256_SHA256,
+    SignatureScheme::ECDSA_NISTP384_SHA384,
+    SignatureScheme::ECDSA_NISTP521_SHA512,
+    SignatureScheme::ED25519,
+];
+pub const TLS12_RSA_SCHEMES: [SignatureScheme; 6] = [
+    SignatureScheme::RSA_PKCS1_SHA256,
+    SignatureScheme::RSA_PKCS1_SHA384,
+    SignatureScheme::RSA_PKCS1_SHA512,
+    SignatureScheme::RSA_PSS_SHA256,
+    SignatureScheme::RSA_PSS_SHA384,
+    SignatureScheme::RSA_PSS_SHA512,
+];
